fix(agent): session-cumulative cost priced per answering model
The wire's session_cost / session token figures were TURN-scoped
(LoopTurnState resets per user message) and re-priced the whole turn
at whichever model produced the LATEST response:

- after every turn the displayed "session" spend reset to zero;
- a mid-turn provider failover re-priced earlier iterations at the
  new model's rate;
- once profile/llm/select lands, every model switch evicts the
  session runtime — and with it any in-memory notion of prior spend.

Fixes, layered:

- octos-llm: deterministic catalog_pricing substring fallback
  (longest family key wins, then shortest super-key; lexicographic
  tie-break). The old first-HashMap-hit scan handed a random
  sibling's pricing per process to ids matching several keys.
- octos-agent: LoopTurnState::record_usage now takes the response's
  estimated cost priced at the model that ACTUALLY produced it
  (failover slot resolved via provider_index; the verifier priced at
  its own model; tool-reported usage at the active slot). New
  session_usage module: a SharedSessionUsage handle the session
  actor seeds from the durable usage ledger and folds each completed
  run into; emit_cost_update emits base + live turn. CostUpdate
  session token fields widen to u64.
- octos-cli: the session actor owns the handle (injected via
  with_session_usage_base), hydrates it from
  PersistentUsageLedger::session_totals at run() start — surviving
  the runtime-cache eviction a model switch triggers — and folds in
  record_usage_event with the same numbers the ledger records, so
  re-seeding reproduces exactly the sum of folds. The _completion
  footer's session_cost now reads the folded handle (was: this turn
  re-priced at the final model).

TUI/web already render these fields as session stats; they now
receive what the names promised.

+++ crates/octos-agent/src/progress.rs
@@ -144,11 +144,16 @@ pub enum ProgressEvent {
 
     /// Cost update after a response.
     CostUpdate {
-        session_input_tokens: u32,
-        session_output_tokens: u32,
+        /// Session-cumulative token counts: completed runs folded into the
+        /// shared session-usage base (when the embedder injected one) plus
+        /// the live turn. u64 because sessions outlive single turns.
+        session_input_tokens: u64,
+        session_output_tokens: u64,
         /// Cost of this response (None if pricing unknown).
         response_cost: Option<f64>,
-        /// Cumulative session cost.
+        /// Cumulative session cost: each completed run and each response
+        /// of the live turn priced at the model that produced it. `None`
+        /// until any priced usage exists.
         session_cost: Option<f64>,
         /// Model identifier that produced this response. Forwarded by the
         /// API bridge into `metadata.token_cost.model` so chat clients can
